nlp-sql-v2/requirements.txt
# Core FastAPI
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Vanna AI for text-to-SQL
vanna==0.8.0

# LLM - Anthropic Claude
anthropic==0.40.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
